requirements.txt
llama-index
transformers
python-docx
openai